+++ base/src/main/java/io/github/sds100/keymapper/base/actions/ActionData.kt
@@ -383,12 +383,18 @@ sealed class ActionData : Comparable<ActionData> {
         }
 
         @Serializable
-        data class StepForward(override val packageName: String) : ControlMediaForApp() {
+        data class StepForward(
+            override val packageName: String,
+            val stepDurationMs: Long? = null,
+        ) : ControlMediaForApp() {
             override val id = ActionId.STEP_FORWARD_PACKAGE
         }
 
         @Serializable
-        data class StepBackward(override val packageName: String) : ControlMediaForApp() {
+        data class StepBackward(
+            override val packageName: String,
+            val stepDurationMs: Long? = null,
+        ) : ControlMediaForApp() {
             override val id = ActionId.STEP_BACKWARD_PACKAGE
         }
     }
@@ -436,12 +442,12 @@ sealed class ActionData : Comparable<ActionData> {
         }
 
         @Serializable
-        data object StepForward : ControlMedia() {
+        data class StepForward(val stepDurationMs: Long? = null) : ControlMedia() {
             override val id = ActionId.STEP_FORWARD
         }
 
         @Serializable
-        data object StepBackward : ControlMedia() {
+        data class StepBackward(val stepDurationMs: Long? = null) : ControlMedia() {
             override val id = ActionId.STEP_BACKWARD
         }
     }