+++ base/src/main/java/io/github/sds100/keymapper/base/actions/PerformActionsUseCase.kt
@@ -243,11 +243,11 @@ class PerformActionsUseCaseImpl @AssistedInject constructor(
             }
 
             is ActionData.ControlMediaForApp.StepForward -> {
-                result = mediaAdapter.stepForward(action.packageName)
+                result = mediaAdapter.stepForward(action.packageName, action.stepDurationMs)
             }
 
             is ActionData.ControlMediaForApp.StepBackward -> {
-                result = mediaAdapter.stepBackward(action.packageName)
+                result = mediaAdapter.stepBackward(action.packageName, action.stepDurationMs)
             }
 
             is ActionData.Rotation.CycleRotations -> {
@@ -651,11 +651,11 @@ class PerformActionsUseCaseImpl @AssistedInject constructor(
             }
 
             is ActionData.ControlMedia.StepForward -> {
-                result = mediaAdapter.stepForward()
+                result = mediaAdapter.stepForward(durationMs = action.stepDurationMs)
             }
 
             is ActionData.ControlMedia.StepBackward -> {
-                result = mediaAdapter.stepBackward()
+                result = mediaAdapter.stepBackward(durationMs = action.stepDurationMs)
             }
 
             is ActionData.GoBack -> {